Updated script with different noise mechanics
Now we use noisy signal only when giving the agent the reward, and we log only noiseless data.
Thus when we look at misalignment it is the misalignment between only the beams chosen by the agent and not between the actual rewards received.

We also added a value for the noise power in the multi agent script, based on suggestion by supervisor.